fix: rename event class, apply requested changes
